# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlowForge CMS API",
    "version": "1.0.0",
    "description": "Multi-tenant headless CMS API. Manage content types, entries, media, webhooks, and more.",
    "contact": { "name": "FlowForge" },
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "/api/v1", "description": "API v1" }
  ],
  "security": [
    { "bearerAuth": [] },
    { "apiKeyAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from /api/v1/auth/login or Clerk session token"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key created in the dashboard"
      }
    },
    "schemas": {
      "ContentType": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "fields": { "type": "array", "items": { "$ref": "#/components/schemas/FieldDefinition" } },
          "locales": { "type": "array", "items": { "type": "string" } },
          "cacheTTL": { "type": "integer" },
          "workflowEnabled": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "FieldDefinition": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["String", "Number", "Date", "Boolean", "RichText", "Reference"] },
          "required": { "type": "boolean" },
          "localizable": { "type": "boolean" },
          "refContentType": { "type": "string" },
          "defaultValue": { "type": "string" }
        }
      },
...
      "Media": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "mimeType": { "type": "string" },
          "size": { "type": "integer" },
          "url": { "type": "string" },
          "type": { "type": "string", "enum": ["image", "document", "video", "audio", "other"] }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": { "type": "array", "items": { "type": "object" } }
        }
      }
    }
  },
  "paths": {
...
    "/sitemap.xml": {
      "get": {
        "summary": "Generate sitemap",
        "tags": ["System"],
        "parameters": [
          { "name": "tenant", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "XML sitemap" }
        }
      }
    },
    "/content-types": {
      "get": {
        "summary": "List all content types",
        "tags": ["Content Types"],
        "responses": {
          "200": { "description": "Array of content types", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ContentType" } } } } }
        }
      },
      "post": {
        "summary": "Create a content type",
        "tags": ["Content Types"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContentType" } } } },
        "responses": {
          "201": { "description": "Created content type" },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/media": {
      "get": {
        "summary": "List media",
        "tags": ["Media"],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["image", "document", "video", "audio", "other"] } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Paginated media", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Media" } } } } } }
      },
      "post": {
        "summary": "Upload media",
        "tags": ["Media"],
        "requestBody": { "required": true, "content": { "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" } } } } } },
        "responses": { "201": { "description": "Uploaded media" } }
      }
    },
...
    "/graphql": {
      "post": {
        "summary": "GraphQL endpoint",
        "tags": ["GraphQL"],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "query": { "type": "string" }, "variables": { "type": "object" } } } } } },
        "responses": { "200": { "description": "GraphQL response" } }
      }
    },
    "/search": {
      "get": {
        "summary": "Full-text search across entries",
        "tags": ["Search"],
        "parameters": [{ "name": "q", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Search results" } }
      }
    },
...
    "/audit-logs": {
      "get": { "summary": "List audit logs", "tags": ["Audit Logs"], "parameters": [{ "name": "page", "in": "query", "schema": { "type": "integer" } }, { "name": "limit", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "Paginated audit logs" } } }
    },
...
    "/calendar": {
      "get": { "summary": "Get content calendar", "tags": ["Calendar"], "parameters": [{ "name": "year", "in": "query", "schema": { "type": "integer" } }, { "name": "month", "in": "query", "schema": { "type": "integer" } }], "responses": { "200": { "description": "Calendar data" } } }
    },
...
  },
...
}